In README, make the precondition-checking solution more noticeable
The lack of a given solution to the problem “the caller needs to check that there is a match” almost scared me away from this library. I only found the solution by looking into the code of fzy-demo – it didn’t occur to me to look for the solution within the full example until after I had started trying to edit this README to add the relevant part of fzy-demo’s code. This sentence right after the warning should make it clear that the precondition is not as hard to satisfy as it sounds.